requirements.txt
openai>=1.0
pandas
numpy<2
wikipedia
chardet
openai-whisper
youtube-transcript-api
langgraph>=0.0.34
gradio==4.44.1
huggingface-hub
requests
tiktoken
python-magic
openpyxl
tabulate
langchain
openai-whisper
requests
python-dotenv